No longer use celery for sending notifications
