+++ wagtail/wagtailadmin/views/pages.py
@@ -15,7 +15,8 @@ from django.db.models import Count
 
 from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList
 from wagtail.wagtailadmin.forms import SearchForm, CopyForm
-from wagtail.wagtailadmin import tasks, signals
+from wagtail.wagtailadmin.utils import send_notification
+from wagtail.wagtailadmin import signals
 
 from wagtail.wagtailcore import hooks
 from wagtail.wagtailcore.models import Page, PageRevision, get_navigation_menu_items
@@ -225,7 +226,7 @@ def create(request, content_type_app_name, content_type_model_name, parent_page_
                 messages.success(request, _("Page '{0}' published.").format(page.title))
             elif is_submitting:
                 messages.success(request, _("Page '{0}' submitted for moderation.").format(page.title))
-                tasks.send_notification.delay(page.get_latest_revision().id, 'submitted', request.user.id)
+                send_notification(page.get_latest_revision().id, 'submitted', request.user.id)
             else:
                 messages.success(request, _("Page '{0}' created.").format(page.title))
 
@@ -361,7 +362,7 @@ def edit(request, page_id):
                     messages.button(reverse('wagtailadmin_pages_view_draft', args=(page_id,)), _('View draft')),
                     messages.button(reverse('wagtailadmin_pages_edit', args=(page_id,)), _('Edit'))
                 ])
-                tasks.send_notification.delay(page.get_latest_revision().id, 'submitted', request.user.id)
+                send_notification(page.get_latest_revision().id, 'submitted', request.user.id)
             else:
                 messages.success(request, _("Page '{0}' updated.").format(page.title))
 
@@ -782,7 +783,7 @@ def approve_moderation(request, revision_id):
     if request.method == 'POST':
         revision.approve_moderation()
         messages.success(request, _("Page '{0}' published.").format(revision.page.title))
-        tasks.send_notification.delay(revision.id, 'approved', request.user.id)
+        send_notification(revision.id, 'approved', request.user.id)
 
     return redirect('wagtailadmin_home')
 
@@ -799,7 +800,7 @@ def reject_moderation(request, revision_id):
     if request.method == 'POST':
         revision.reject_moderation()
         messages.success(request, _("Page '{0}' rejected for publication.").format(revision.page.title))
-        tasks.send_notification.delay(revision.id, 'rejected', request.user.id)
+        send_notification(revision.id, 'rejected', request.user.id)
 
     return redirect('wagtailadmin_home')
